requirements.txt
python-dotenv
Langchain
langchain-openai
streamlit
langsmith
langserve
fastapi
uvicorn
sse_starlette
requests
langchain-community
bs4 
pypdf
chromadb
faiss-cpu
arxiv
wikipedia
langchainhub
groq
langchain-groq
langchain_ollama
sentence-transformers
urllib3
langchain_objectbox
langchain_core
langchain-mcp-adapters
mcp